Fix a couple bugs with the `serve` and `build_release` commands (#6659)

+++ tool/lib/commands/build_release.dart
@@ -61,7 +61,9 @@ class BuildReleaseCommand extends Command {
     }
 
     logStatus('cleaning project');
-    webBuildDir.deleteSync(recursive: true);
+    if (webBuildDir.existsSync()) {
+      webBuildDir.deleteSync(recursive: true);
+    }
     await processManager.runProcess(
       CliCommand.flutter('clean'),
       workingDirectory: repo.devtoolsAppDirectoryPath,
